first port to RISC32

+++ manuscript/code/bitwise.c
@@ -2,19 +2,19 @@
 
 // libcstar procedures for printing
 void init_library();
-void print(uint64_t* s);
-void print_integer(uint64_t n);
-void print_binary(uint64_t n, uint64_t a);
+void print(uint32_t* s);
+void print_integer(uint32_t n);
+void print_binary(uint32_t n, uint32_t a);
 void println();
 
 // libcstar procedures for left and right shifting
-uint64_t left_shift(uint64_t n, uint64_t b);
-uint64_t right_shift(uint64_t n, uint64_t b);
+uint32_t left_shift(uint32_t n, uint32_t b);
+uint32_t right_shift(uint32_t n, uint32_t b);
 
-uint64_t main() {
-  uint64_t i;
-  uint64_t j;
-  uint64_t u;
+uint32_t main() {
+  uint32_t i;
+  uint32_t j;
+  uint32_t u;
 
   // initialize selfie's libcstar library
   init_library();
